update load photos

+++ src/main/java/com/wzlibs/androidstoragefiles/photo/model/PhotoFile.kt
@@ -1,3 +1,3 @@
 package com.wzlibs.androidstoragefiles.photo.model
 
-data class PhotoFile(val id: Int, val name: String, val url: String, val dateModifier: Long, val size: Long)
+data class PhotoFile(val id: Long, val name: String, val url: String, val dateModifier: Long, val size: Long)
